backend/requirements.txt
# VideoRAG - Python dependencies
# CPU-only, production-ready

# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3

# ─── Video Download ───────────────────────────────────────────────────────────
yt-dlp==2024.12.13

# ─── Transcription (CPU-optimized) ───────────────────────────────────────────
faster-whisper==1.1.0
ctranslate2==4.5.0

# ─── Embeddings ───────────────────────────────────────────────────────────────
sentence-transformers==3.3.1
transformers==4.47.1
torch==2.5.1+cpu          # CPU-only PyTorch
torchvision==0.20.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# ─── Vector Database ─────────────────────────────────────────────────────────
faiss-cpu==1.9.0
numpy==1.26.4

# ─── LLM (Gemini) ────────────────────────────────────────────────────────────
google-generativeai==0.8.3

# ─── TTS ─────────────────────────────────────────────────────────────────────
kokoro==0.9.2
soundfile==0.12.1
sounddevice==0.5.1

# ─── Fallback TTS ────────────────────────────────────────────────────────────
pyttsx3==2.90

# ─── Audio Processing ─────────────────────────────────────────────────────────
librosa==0.10.2

# ─── Utilities ───────────────────────────────────────────────────────────────
requests==2.32.3
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1
